# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manajemen Proker API",
        "description": "Dokumentasi API untuk Sistem Manajemen Program Kerja. Mencakup Authentication, Menu Management, Role & Group Management, dan Permission Management.",
        "contact": {
            "email": "admin@manajemenproker.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/sys-users": {
            "get": {
                "tags": [
                    "System User"
                ],
                "summary": "Daftar User (Paginated)",
                "description": "Mengambil daftar seluruh user dengan pagination.",
                "operationId": "710d01e8180f9e9a7c538ed548de46ef",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Halaman aktif",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pagesize",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "System User"
                ],
                "summary": "Tambah User Baru",
                "description": "Membuat user baru.",
                "operationId": "6629b5c85af19492eb0663bcf8772730",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "telegram_number": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "id_sys_group": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sys-users/{id}": {
            "get": {
                "tags": [
                    "System User"
                ],
                "summary": "Detail User",
                "description": "Menampilkan detail satu user berdasarkan ID.",
                "operationId": "21502c8d440609ac00c3ef639b5891a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "System User"
                ],
                "summary": "Update User",
                "description": "Memperbarui data user yang sudah ada.",
                "operationId": "529cb2a6902f2be5028ff47cd9a357e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "telegram_number": {
                                        "type": "string"
                                    },
                                    "id_sys_group": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "System User"
                ],
                "summary": "Hapus User",
                "description": "Menghapus user berdasarkan ID.",
                "operationId": "b38492332b5a6fac7139c9d12c88f25f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Masukkan token dengan format: Bearer {access_token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}